Change integer to float type

+++ database/migrations/2016_09_16_010741_alter_table_programs_add_now_price.php
@@ -13,7 +13,7 @@ class AlterTableProgramsAddNowPrice extends Migration
     public function up()
     {
         Schema::table('programs', function ($table) {
-            $table->integer('now_price')->default(100)->after('name');
+            $table->float('now_price')->default(100)->after('name');
         });
     }
